revisi lagi

+++ app/models/artikel_model.php
@@ -67,7 +67,6 @@ class artikel_model
         $this->db->bind("id_kategori", $data['kategori']);
         $this->db->bind("id_artikel", $lastInsertId);
 
-        $this->db->execute();
         return $this->db->rowCount();
     }
 
@@ -99,7 +98,6 @@ class artikel_model
             $this->db->bind("gambar", $data['image']);
         }
         $this->db->bind("id_artikel", $data['id_artikel']);
-        $this->db->execute();
         $jumlah_artikel_diubah = $this->db->rowCount();
 
         $query = "UPDATE kontributor
@@ -109,7 +107,6 @@ class artikel_model
         $this->db->bind("id_kategori", $data['kategori']);
         $this->db->bind("id_artikel", $data['id_artikel']);
         $this->db->bind("id_kontributor", $data['id_kontributor']);
-        $this->db->execute();
         $jumlah_kontributor_diubah = $this->db->rowCount();
 
         if ($jumlah_artikel_diubah > 0) {
@@ -124,14 +121,13 @@ class artikel_model
         $query = "DELETE FROM kontributor WHERE id_kontributor = :id_kontributor";
         $this->db->query($query);
         $this->db->bind("id_kontributor", $id_kontributor);
-        $this->db->execute();
-        $this->db->rowCount();
 
-        $query = "DELETE FROM artikel WHERE id_artikel = :id_artikel";
-        $this->db->query($query);
-        $this->db->bind("id_artikel", $id_artikel);
-        $this->db->execute();
+        if ($this->db->rowCount() > 0) {
+            $query = "DELETE FROM artikel WHERE id_artikel = :id_artikel";
+            $this->db->query($query);
+            $this->db->bind("id_artikel", $id_artikel);
 
-        return $this->db->rowCount();
+            return $this->db->rowCount();
+        }
     }
 }
